Mv network work to Data
