+++ app/src/main/java/ru/surf/learn2invest/ui/components/screens/fragments/marketreview/MarketReviewAdapter.kt
@@ -13,10 +13,10 @@ import coil.ImageLoader
 import coil.load
 import dagger.hilt.android.qualifiers.ActivityContext
 import ru.surf.learn2invest.R
-import ru.surf.learn2invest.noui.network_components.responses.CoinReviewResponse
+import ru.surf.learn2invest.data.network_components.responses.CoinReviewResponse
 import ru.surf.learn2invest.ui.components.screens.fragments.asset_review.AssetReviewActivity
 import ru.surf.learn2invest.utils.AssetConstants
-import ru.surf.learn2invest.utils.RetrofitLinks.API_ICON
+import ru.surf.learn2invest.data.network_components.RetrofitLinks.API_ICON
 import ru.surf.learn2invest.utils.getWithCurrency
 import ru.surf.learn2invest.utils.round
 import java.text.NumberFormat
