+++ app/src/main/java/ru/surf/learn2invest/ui/components/screens/fragments/subhistory/SubHistoryAdapter.kt
@@ -14,7 +14,7 @@ import dagger.hilt.android.qualifiers.ActivityContext
 import ru.surf.learn2invest.R
 import ru.surf.learn2invest.data.database_components.entity.transaction.Transaction
 import ru.surf.learn2invest.domain.TransactionsType
-import ru.surf.learn2invest.utils.RetrofitLinks.API_ICON
+import ru.surf.learn2invest.data.network_components.RetrofitLinks.API_ICON
 import javax.inject.Inject
 
 class SubHistoryAdapter @Inject constructor(
