+++ data/src/main/java/ru/surf/learn2invest/data/network_components/responses/APIWrapper.kt
@@ -1,4 +1,4 @@
-package ru.surf.learn2invest.noui.network_components.responses
+package ru.surf.learn2invest.data.network_components.responses
 
 /**
  * Обертка верхнего уровня для парсинга JSON
