+++ data/src/main/java/ru/surf/learn2invest/data/network_components/responses/CoinReviewResponse.kt
@@ -1,9 +1,9 @@
-package ru.surf.learn2invest.noui.network_components.responses
+package ru.surf.learn2invest.data.network_components.responses
 
 /**
- * Структура JSON для парсинга подробной информации об активе
+ * Структура JSON для парсинга иформации о рынке (элемент массива)
  */
-data class AugmentedCoinReviewResponse(
+data class CoinReviewResponse(
     val id: String, // идентификатор коина в API
     val rank: Int,
     val symbol: String, // аббревиатура
@@ -14,6 +14,5 @@ data class AugmentedCoinReviewResponse(
     val volumeUsd24Hr: Float,
     val priceUsd: Float,
     val changePercent24Hr: Float,
-    val vwap24Hr: Float,
-    val explorer: String
+    val vwap24Hr: Float
 )
